@@ -84,17 +84,17 @@
     "gatsby": "^2.21.21",
     "gh-pages": "^1.1.0",
     "jest": "^25.2.4",
-    "metro-react-native-babel-preset": "^0.53.1",
+    "metro-react-native-babel-preset": "^0.58.0",
     "mocha": "^5.0.0",
     "prettier-eslint": "^9.0.0",
-    "react": "16.9.0",
+    "react": "16.11.0",
     "react-autobind": "^1.0.6",
     "react-dom": "^15.4.2",
     "react-js-to-ts": "^1.4.0",
-    "react-native": "0.61.5",
+    "react-native": "0.62.2",
     "react-native-gesture-handler": "1.5.6",
     "react-native-keyboard-tracking-view": "^5.6.1",
-    "react-native-navigation": "^3.4.0",
+    "react-native-navigation": "^6.4.0",
     "react-native-reanimated": "^1.7.0",
     "react-test-renderer": "^16.13.1",
     "shell-utils": "^1.0.10",
