@@ -1,8 +1,20 @@
 import React, { PureComponent } from 'react';
-import { LayoutChangeEvent, ImageStyle, TextStyle } from 'react-native';
+import { LayoutChangeEvent, ImageStyle, TextStyle, StyleProp } from 'react-native';
 import { BaseComponentInjectedProps, ForwardRefInjectedProps, TypographyModifiers, ColorsModifiers, BackgroundColorModifier, MarginModifiers } from '../../commons/new';
+import { TouchableOpacityProps } from '../touchableOpacity';
 import { TextPropTypes } from '../text';
-export declare type ButtonPropTypes = TextPropTypes & TypographyModifiers & ColorsModifiers & BackgroundColorModifier & MarginModifiers & {
+export declare enum ButtonSize {
+    xSmall = "xSmall",
+    small = "small",
+    medium = "medium",
+    large = "large"
+}
+export declare enum AnimationDirection {
+    center = "center",
+    left = "left",
+    right = "right"
+}
+export declare type ButtonPropTypes = TouchableOpacityProps & TypographyModifiers & ColorsModifiers & BackgroundColorModifier & MarginModifiers & {
     /**
      * Text to show inside the button
      */
@@ -18,7 +30,7 @@ export declare type ButtonPropTypes = TextPropTypes & TypographyModifiers & Colo
     /**
      * Icon image style
      */
-    iconStyle?: ImageStyle;
+    iconStyle?: StyleProp<ImageStyle>;
     /**
      * Should the icon be right to the label
      */
@@ -30,7 +42,7 @@ export declare type ButtonPropTypes = TextPropTypes & TypographyModifiers & Colo
     /**
      * Size of the button [large, medium, small, xSmall]
      */
-    size?: 'xSmall' | 'small' | 'medium' | 'large';
+    size?: ButtonSize;
     /**
      * Custom border radius.
      */
@@ -70,7 +82,7 @@ export declare type ButtonPropTypes = TextPropTypes & TypographyModifiers & Colo
     /**
      * Props that will be passed to the button's Text label.
      */
-    labelProps?: object;
+    labelProps?: TextPropTypes;
     /**
      * should the button act as a coast to coast button (no border radius)
      */
@@ -104,7 +116,7 @@ export declare type ButtonPropTypes = TextPropTypes & TypographyModifiers & Colo
     /**
      * the direction of the animation ('left' and 'right' will effect the button's own alignment)
      */
-    animateTo?: 'center' | 'left' | 'right';
+    animateTo?: AnimationDirection;
 };
 export declare type ButtonState = {
     size?: number;
@@ -125,17 +137,8 @@ declare class Button extends PureComponent<Props, ButtonState> {
     static defaultProps: {
         iconOnRight: boolean;
     };
-    static sizes: {
-        xSmall: string;
-        small: string;
-        medium: string;
-        large: string;
-    };
-    static animationDirection: {
-        center: string;
-        left: string;
-        right: string;
-    };
+    static sizes: typeof ButtonSize;
+    static animationDirection: typeof AnimationDirection;
     constructor(props: Props);
     state: {
         size: undefined;
@@ -316,7 +319,7 @@ declare class Button extends PureComponent<Props, ButtonState> {
     } | {
         shadowColor: any;
     })[] | undefined;
-    getIconStyle(): (ImageStyle | undefined)[];
+    getIconStyle(): StyleProp<ImageStyle>[];
     getAnimationDirectionStyle(): {
         alignSelf: string;
     } | undefined;
@@ -325,5 +328,5 @@ declare class Button extends PureComponent<Props, ButtonState> {
     render(): JSX.Element;
 }
 export { Button };
-declare const _default: React.ComponentClass<ButtonPropTypes, any> | React.FunctionComponent<ButtonPropTypes>;
+declare const _default: React.ComponentClass<ButtonPropTypes, any> & typeof Button;
 export default _default;
