@@ -4,24 +4,26 @@
  * Please use this file for declaring all the exports, so they could be picked up by typescript's complier
  */
 export * from './style';
-export {withScrollEnabler} from './commons/new';
+export {asBaseComponent, withScrollEnabler, withScrollReached, WithScrollEnablerProps, WithScrollReachedProps} from './commons/new';
+export {default as Avatar, AvatarPropTypes} from './components/avatar';
 export {default as Card, CardPropTypes, CardSectionProps} from './components/card';
 export {default as View, ViewPropTypes} from './components/view';
-export {default as Text} from './components/text';
+export {default as Text, TextPropTypes} from './components/text';
 export {default as TouchableOpacity, TouchableOpacityProps} from './components/touchableOpacity';
-export {default as Button} from './components/button';
-export {default as Checkbox} from './components/checkbox';
-export {default as Image} from './components/image';
-export {default as Overlay} from './components/overlay';
-export {default as RadioButton} from './components/radioButton/RadioButton';
-export {default as RadioGroup} from './components/radioButton/RadioGroup';
+export {default as Button, ButtonPropTypes} from './components/button';
+export {default as Checkbox, CheckboxPropTypes} from './components/checkbox';
+export {default as Image, ImageProps} from './components/image';
+export {default as Overlay, OverlayTypes} from './components/overlay';
+export {default as RadioButton, RadioButtonPropTypes} from './components/radioButton/RadioButton';
+export {default as RadioGroup, RadioGroupPropTypes} from './components/radioButton/RadioGroup';
+export {default as TabBar} from './components/TabBar';
 
 /* All components with manual typings */
 export {
   ActionBar,
   ActionSheet,
-  Avatar,
   Badge,
+  BadgeProps,
   Card,
   Carousel,
   ConnectionStatusBar,
@@ -45,7 +47,6 @@ export {
   GradientSlider,
   ColorSliderGroup,
   Stepper,
-  TabBar,
   TagsInput,
   SharedTransition,
   StackAggregator,
@@ -63,8 +64,11 @@ export {
   Modal,
   StateScreen,
   WheelPicker,
+  WheelPickerProps,
   Incubator,
-  ColorPicker
+  ColorPicker,
+  Picker,
+  PickerProps
 } from '../typings';
 
 /* All components that are missing either manual or auto generated typings */
@@ -77,7 +81,6 @@ export const HighlighterOverlayView;
 export const Keyboard;
 export const KeyboardAwareListView;
 export const KeyboardAwareScrollView;
-export const Picker;
 export const SafeAreaInsetsManager;
 export const SafeAreaSpacerView;
 export const ScrollBar;
