@@ -29,8 +29,8 @@ export declare class ThemeManager {
     };
     setItem(key: string, value: string): void;
     getItem(key: string): any;
-    setComponentTheme(componentName: string, overrides: Dictionary<string | number> | Function): void;
-    setComponentForcedTheme(componentName: string, overrides: Dictionary<string> | Function): void;
+    setComponentTheme(componentName: string, overrides: Dictionary<any> | Function): void;
+    setComponentForcedTheme(componentName: string, overrides: Dictionary<any> | Function): void;
     get components(): Extendable;
     get forcedThemeComponents(): Extendable;
     get primaryColor(): string;
