@@ -1,5 +1,5 @@
 //
-//  RCTCustomKeyboardViewController.h
+//  RCTCustomKeyboardViewControllerTemp.h
 //
 //  Created by Leo Natan (Wix) on 12/12/2016.
 //  Copyright © 2016 Leo Natan. All rights reserved.
@@ -13,9 +13,10 @@
 #import "RCTRootView.h"
 #endif
 
-@interface RCTCustomKeyboardViewController : UIInputViewController
+@interface RCTCustomKeyboardViewControllerTemp : UIInputViewController
 
 - (void) setAllowsSelfSizing:(BOOL)allowsSelfSizing;
+- (instancetype)initWithUsingSafeArea:(BOOL)useSafeArea;
 
 @property (nonatomic, strong) NSLayoutConstraint *heightConstraint;
 @property (nonatomic, strong) RCTRootView *rootView;
