@@ -4,7 +4,7 @@ import {BaseComponentPaddingModifier, BaseComponentMarginModifier} from './spaci
 
 type TypographyName = 'fixme';
 type ColorName = 'fixme';
-export class UIComponent extends React.PureComponent {}
+export class UIComponent<P = {}, S = {}, SS = any> extends React.PureComponent<P, S, SS> {}
 export function forwardRef<T, P>(
   WrappedComponent: React.ComponentType<T, P>
 ): React.ComponentType<T, P>;
@@ -43,6 +43,10 @@ export type BaseComponentModifiers =
 export class BaseComponent<P = {}, S = {}, SS = any> extends React.Component<P & BaseComponentModifiers, S, SS> {
   static propTypes?: any;
   static defaultProps?: Partial<P>;
+
+  // TODO: There's more methods here, need to create more precise types
+  getThemeProps: () => P & ThemeProps;
+  extractAccessibilityProps: () => AccessibilityProps;
 }
 
 export type _B = Partial<Record<BaseComponentColorModifierVariations, boolean>>;
@@ -50,5 +54,9 @@ export type _B = Partial<Record<BaseComponentColorModifierVariations, boolean>>;
 export class PureBaseComponent<P = {}, S = {}, SS = any> extends React.PureComponent<P & _B, S, SS> {
   static propTypes?: any;
   static defaultProps?: Partial<P>;
+
+  // TODO: There's more methods here, need to create more precise types
+  getThemeProps: () => P & ThemeProps;
+  extractAccessibilityProps: () => AccessibilityProps;
 }
 
